build_proton: explicitly disable openal examples
This makes the build more deterministic, as whether the examples were
enabled or not depended on the availability of SDL_sound and ffmpeg on
the host, and when it does get enabled, it causes build failures.

We're passing -m32 for gcc, but not for g++, which causes it to fail to
link to the just-built openal objects, and it also tries to link to
a 32-bit avformat/avcodec library on the host, which isn't generally
installed on the mac, and is unlikely elsewhere.

Signed-off-by: Christopher Larson <kergoth@gmail.com>